backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
langchain==0.0.312
langchain-openai==0.0.2
python-dotenv==1.0.0
sse-starlette==1.6.5 